pyproject.toml
[project]
name = "valheim-blueprint"
version = "0.1.0"
description = "Generate Valheim blueprints from natural language using Claude"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.scripts]
valheim-blueprint = "src.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
